Add: 960 grid CSS template
On the design image the gutter on the left is 10px and on right is 13px. The spaces between the sections vary between 13px and 11px. The three small sections at the bottom are 202px wide and the product sidebar is 206px wide.

The background to the main menu is 637px wide. I've made the columns 202px wide and the gutter widths 16px so the background to the main menu would fit nicely in the last three columns (202px * 3 + 16px * 2 = 238px) meaning the vertical alignment might be nice. This does of course mean taking the liberty of moving the main menu right slightly.